% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{FS_5kbLoops.txt}
\alias{FS_5kbLoops.txt}
\title{Example NHA9 (FS) Loops}
\value{
Downloads and caches a \code{.txt} file in BEDPE format.
}
\description{
Loops called with SIP (https://github.com/PouletAxel/SIP)
using Hi-C data from HEK cells expressing an NHA9 fusion
with an F to S mutated IDR.
}
\examples{

bedpeFile <- FS_5kbLoops.txt()
bedpeFile

}
