% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LEUK_HEK_PJA27_inter_30.mcool}
\alias{LEUK_HEK_PJA27_inter_30.mcool}
\title{Example NHA9 (FS) Hi-C data in .mcool format}
\value{
Downloads and caches a \code{.mcool} file.
}
\description{
Subsampled Hi-C in HEK cells expressing the
NHA9 fusion with an F to S mutated IDR,
converted to \code{.mcool}.
}
\examples{

mcoolFile <- LEUK_HEK_PJA27_inter_30.mcool()
mcoolFile

}
