% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LEUK_HEK_PJA30_inter_30.hic}
\alias{LEUK_HEK_PJA30_inter_30.hic}
\title{Example Hi-C data}
\value{
Downloads and caches a \code{.hic} file.
}
\description{
Subsampled Hi-C in HEK cells expressing the
NHA9 fusion with a Wildtype IDR.
}
\examples{

hicFile <- LEUK_HEK_PJA30_inter_30.hic()
hicFile

}
