% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LEUK_HEK_PJA30_inter_30_2500Kb.cool}
\alias{LEUK_HEK_PJA30_inter_30_2500Kb.cool}
\title{Example NHA9 (WT) Hi-C data in .cool format}
\value{
Downloads and caches a \code{.cool} file.
}
\description{
Subsampled Hi-C in HEK cells expressing the
NHA9 fusion with a Wildtype IDR,
converted to \code{.cool} at 2.5Mb resolution.
}
\examples{

coolFile <- LEUK_HEK_PJA30_inter_30_2500Kb.cool()
coolFile

}
