% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LIMA_0240.bedpe}
\alias{LIMA_0240.bedpe}
\title{Example Timecourse Loops}
\value{
Downloads and caches a \code{.bedpe} file in BEDPE format.
}
\description{
Loops called with SIP (https://github.com/PouletAxel/SIP)
using Hi-C data from THP-1 cells exposed to LPS & IF-G
for 240 minutes.
}
\examples{

bedpeFile <- LIMA_0240.bedpe()
bedpeFile

}
